+++ src/main/java/roomescape/auth/AuthCustomAnnotation.java
@@ -1,4 +1,4 @@
-package auth;
+package roomescape.auth;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
