+++ src/main/java/roomescape/auth/AuthRoleInterceptor.java
@@ -1,11 +1,13 @@
-package auth;
+package roomescape.auth;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import jwt.JWTUtils;
 import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.HandlerInterceptor;
 
-
+@Component
 @RequiredArgsConstructor
 public class AuthRoleInterceptor implements HandlerInterceptor {
