+++ src/main/java/roomescape/auth/AuthToken.java
@@ -1,4 +1,4 @@
-package auth;
+package roomescape.auth;
 
 public record AuthToken(
         String token
