+++ src/main/java/roomescape/auth/AuthWebConfig.java
@@ -1,4 +1,4 @@
-package auth;
+package roomescape.auth;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
