+++ src/main/java/roomescape/reservation/ReservationController.java
@@ -1,7 +1,5 @@
 package roomescape.reservation;
 
-import auth.AuthClaims;
-import auth.AuthCustomAnnotation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -10,6 +8,8 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+import roomescape.auth.AuthClaims;
+import roomescape.auth.AuthCustomAnnotation;
 
 import java.net.URI;
 import java.util.List;
