+++ src/main/java/roomescape/waiting/WaitingController.java
@@ -1,7 +1,5 @@
 package roomescape.waiting;
 
-import auth.AuthClaims;
-import auth.AuthCustomAnnotation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -9,6 +7,8 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+import roomescape.auth.AuthClaims;
+import roomescape.auth.AuthCustomAnnotation;
 
 import java.net.URI;
