+++ src/test/java/roomescape/MissionStepTest.java
@@ -1,10 +1,10 @@
 package roomescape;
 
-import auth.JWTUtils;
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
+import jwt.JWTUtils;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.stereotype.Component;
